Make compiler run (migration still disabled)
